refactor(meditation): extracts chart to a separate component
