+++ src/app/meditation/chart/meditation-chart.css
@@ -1,0 +1,11 @@
+.chart {
+  display: inline-block;
+  height: 140px !important;
+  width: 100%;
+}
+
+@media screen and (max-width: 480px) {
+    .chart {
+        height: 100px !important;
+    }
+}
