+++ src/app/meditation/chart/meditation-chart.html
@@ -1,0 +1,12 @@
+<div class="chart-wrapper">
+  <base-chart
+    class="chart"
+    [datasets]="chartData"
+    [labels]="chartLabels"
+    [options]="chartOptions"
+    [colors]="chartColors"
+    [legend]="false"
+    chartType="bar"
+  >
+  </base-chart>
+</div>
